+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2026-08-02]
+### Updated
+- The `BT_LANE_EJECT`, `BT_LANE_MOVE`, and `BT_CHANGE_TOOL` macros will now accept either a lane defined as `lane1` or
+  just a numeric indication of the lane such as `1`. 
 ## [2026-07-26]
 ### Fixed
 - Fixed issue 804 by exposing AFC_VERSION to `afc/status` and AFC object endpoints for moonraker